refactor: change location of Rota
